+++ validateframework/src/main/java/cc/doctor/framework/validate/annotation/Min.java
@@ -12,5 +12,5 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Min {
-    long value();
+    double value();
 }
